requirements.txt
cycler==0.10.0
deap==1.2.2
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.9.1
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.1
opencv-python==4.0.0.21
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
scipy==1.2.1
six==1.12.0
